Add meteor staff item
Add a meteor staff item, which summons meteors from the sky when right-clicking. Players can point to where it should fall to, and it has a default max range of 100 blocks out.

The max use range, blast radius, and block breaking can all be configured.

Fix bugs in the configuration handler that would cause some values to reset to default.

+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/item/projectile/gun/AbstractGunItem.java
@@ -403,7 +403,7 @@ public abstract class AbstractGunItem extends Item implements Vanishable {
 		bulletEntity.setFiringItem(firingItem.getItem());
 
 		bulletEntity.shootFromRotation(player, player.xRot, player.yRot,
-				0.0F, CommonConfig.FLINTLOCK_PISTOL_FIRE_VELOCITY.get(), CommonConfig.FLINTLOCK_PISTOL_FIRE_INACCURACY.get());
+				0.0F, CommonConfig.FLINTLOCK_PISTOL_FIRE_VELOCITY.get().floatValue(), CommonConfig.FLINTLOCK_PISTOL_FIRE_INACCURACY.get().floatValue());
 
 		// Roll for random crits
 		if (GeneralUtilities.getRandomNumber(0f, 1f) <= CommonConfig.GUN_CRIT_CHANCE.get()) {
