+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/item/projectile/gun/MusketItem.java
@@ -34,7 +34,7 @@ public class MusketItem extends AbstractGunItem {
 		bulletEntity.setFiringItem(firingItem.getItem());
 
 		bulletEntity.shootFromRotation(player, player.xRot, player.yRot,
-				0.0F, CommonConfig.MUSKET_FIRE_VELOCITY.get(), CommonConfig.MUSKET_FIRE_INACCURACY.get());
+				0.0F, CommonConfig.MUSKET_FIRE_VELOCITY.get().floatValue(), CommonConfig.MUSKET_FIRE_INACCURACY.get().floatValue());
 
 		// Roll for random crits
 		if (GeneralUtilities.getRandomNumber(0f, 1f) <= CommonConfig.GUN_CRIT_CHANCE.get()) {
